+++ packages/radix-ui-themes/src/components/popover.tsx
@@ -7,9 +7,9 @@ import { popoverContentPropDefs } from './popover.props';
 import { extractProps } from '../helpers';
 import { Theme } from '../theme';
 
-import type { GetPropDefTypes } from '../helpers';
+import type { GetPropDefTypes, PropsWithoutRefOrColor } from '../helpers';
 
-interface PopoverRootProps extends React.ComponentPropsWithoutRef<typeof PopoverPrimitive.Root> {}
+interface PopoverRootProps extends PropsWithoutRefOrColor<typeof PopoverPrimitive.Root> {}
 const PopoverRoot: React.FC<PopoverRootProps> = (props: PopoverRootProps) => (
   <PopoverPrimitive.Root {...props} />
 );
@@ -17,7 +17,7 @@ PopoverRoot.displayName = 'PopoverRoot';
 
 type PopoverTriggerElement = React.ElementRef<typeof PopoverPrimitive.Trigger>;
 interface PopoverTriggerProps
-  extends Omit<React.ComponentPropsWithoutRef<typeof PopoverPrimitive.Trigger>, 'asChild'> {}
+  extends Omit<PropsWithoutRefOrColor<typeof PopoverPrimitive.Trigger>, 'asChild'> {}
 const PopoverTrigger = React.forwardRef<PopoverTriggerElement, PopoverTriggerProps>(
   (props, forwardedRef) => <PopoverPrimitive.Trigger {...props} ref={forwardedRef} asChild />
 );
@@ -26,7 +26,7 @@ PopoverTrigger.displayName = 'PopoverTrigger';
 type PopoverContentElement = React.ElementRef<typeof PopoverPrimitive.Content>;
 type PopoverContentOwnProps = GetPropDefTypes<typeof popoverContentPropDefs>;
 interface PopoverContentProps
-  extends Omit<React.ComponentPropsWithoutRef<typeof PopoverPrimitive.Content>, 'asChild'>,
+  extends Omit<PropsWithoutRefOrColor<typeof PopoverPrimitive.Content>, 'asChild'>,
     PopoverContentOwnProps {
   container?: React.ComponentProps<typeof PopoverPrimitive.Portal>['container'];
 }
@@ -56,7 +56,7 @@ PopoverContent.displayName = 'PopoverContent';
 
 type PopoverCloseElement = React.ElementRef<typeof PopoverPrimitive.Close>;
 interface PopoverCloseProps
-  extends Omit<React.ComponentPropsWithoutRef<typeof PopoverPrimitive.Close>, 'asChild'> {}
+  extends Omit<PropsWithoutRefOrColor<typeof PopoverPrimitive.Close>, 'asChild'> {}
 const PopoverClose = React.forwardRef<PopoverCloseElement, PopoverCloseProps>(
   (props, forwardedRef) => <PopoverPrimitive.Close {...props} ref={forwardedRef} asChild />
 );